Fix desktop rules and moving background when sen scroll
Add function that change the header background position everytime a scroll event in the header element is activated.

Work in progress, need to find a way that sets the background position accurately when the header element is scrolled all the way to the top. currently the background position value plays between 0.3px to 0.12px when the header element is scrolled all the way to the top "